Get operations

This function returns information about operations

Usage

get_operations(
  code = NULL,
  resource = "all",
  help = FALSE,
  ioe = FALSE,
  geographical_granularity = NULL,
  temporal_granularity = NULL,
  verbose = TRUE,
  lang = "ES"
)

Arguments

code

(string) operation identificator

resource

(string) resource to access, by default resource = "metadata" to get serie metadata. Possible values are all, metadata, variables_values or by_granularity

help

(boolean) type any value for resource param and type help = TRUE to see params available for this resource.

ioe

(boolean) TRUE if code is in format IO30138, and FALSE by default

geographical_granularity

(string) geographical granularity

temporal_granularity

(string) temporal granularity

verbose

(boolean) show more information during the process

lang

(string) language used to obtain information

Examples

get_operations()
#>     Id Cod_IOE
#> 1    4   30147
#> 2    6   30211
#> 3    7   30168
#> 4   10   30256
#> 5   13   30219
#> 6   14   30182
#> 7   15   30457
#> 8   16        
#> 9   18   30180
#> 10  20   30013
#> 11  21   30259
#> 12  22   30245
#> 13  23   30417
#> 14  24   30048
#> 15  25   30138
#> 16  26   30050
#> 17  27   30051
#> 18  30        
#> 19  31   30183
#> 20  32   30103
#> 21  33   30264
#> 22  35        
#> 23  36        
#> 24  39   30162
#> 25  40   30149
#> 26  41   30053
#> 27  42   30052
#> 28  43   30203
#> 29  44   30198
#> 30  48   30071
#> 31  49        
#> 32  51   30019
#> 33  52   30269
#> 34  54   30270
#> 35  61   30222
#> 36  62   30223
#> 37  63   30229
#> 38  66   30463
#> 39  71   30277
#> 40  72   30321
#> 41  93   30049
#> 42 104   30062
#> 43 121   30133
#> 44 125   30151
#> 45 130   30177
#> 46 132   30179
#> 47 137   30185
#> 48 139   30188
#> 49 140   30189
#> 50 155   30453
#> 51 163   30199
#> 52 171   30221
#> 53 180   30230
#> 54 197   30271
#> 55 205   30310
#> 56 212   30465
#> 57 213   30466
#> 58 214   30467
#> 59 215   30468
#> 60 234   30209
#> 61 237   30024
#> 62 238   30235
#> 63 239   30236
#> 64 240   30237
#> 65 241   30238
#> 66 246   30026
#> 67 247   30023
#> 68 249   30063
#> 69 254   30220
#> 70 256   30232
#> 71 258   30083
#> 72 259   30153
#> 73 293   30308
#> 74 297   30163
#> 75 303   30187
#> 76 305   30302
#> 77 307   30304
#> 78 309   30306
#> 79 314   30458
#> 80 328   30239
#> 81 329   16029
#> 82 330   16028
#> 83 331        
#> 84 333   30456
#> 85 334   16023
#> 86 336   30279
#> 87 345   30030
#> 88 353        
#> 89 402        
#> 90 406   30324
#> 91 407        
#> 92 410        
#> 93 412        
#> 94 424   30323
#> 95 425        
#> 96 426        
#> 97 429        
#> 98 432        
#> 99 433        
#>                                                                                                Nombre
#> 1                                                        Estadística de Efectos de Comercio Impagados
#> 2                                                                  Índice de Coste Laboral Armonizado
#> 3                                              Estadística de Transmisión de Derechos de la Propiedad
#> 4                                                                                 Indicadores Urbanos
#> 5                                                             Estadística del Procedimiento Concursal
#> 6                                                             Índices de Precios del Sector Servicios
#> 7                                                              Índice de Precios de la Vivienda (IPV)
#> 8                                                                             Distribución de Nombres
#> 9                                                      Índice de Precios de Consumo Armonizado (IPCA)
#> 10                                              Contabilidad Nacional Trimestral de España. Base 2000
#> 11                                                        Estimaciones de la Población Actual (ePOBa)
#> 12           Cifras Oficiales de Población de los Municipios Españoles: Revisión del Padrón Municipal
#> 13                                                Estadística de Defunciones según la Causa de Muerte
#> 14                                             Estadística Estructural de Empresas: Sector Industrial
#> 15                                                                 Índice de Precios de Consumo (IPC)
#> 16                                                                   Índices de Producción Industrial
#> 17                                                                    Índices de Precios Industriales
#> 18                                          Encuesta de coyuntura de comercio al por menor. Base 1994
#> 19                                                      Indicadores de Actividad del Sector Servicios
#> 20                                                                   Índices de Comercio al por Menor
#> 21                                                                   Indicadores Demográficos Básicos
#> 22    Poblaciones de hecho desde 1900 hasta 1991. Cifras oficiales sacadas de los Censos respectivos.
#> 23                 Poblaciones de derecho desde 1986 hasta 1995. Cifras oficiales sacadas del Padrón.
#> 24                                                   Encuesta de Ocupación en Alojamientos Turísticos
#> 25                                                                           Estadística de Hipotecas
#> 26                                                                     Índices de Entradas de Pedidos
#> 27                                                      Índices de Cifras de Negocios en la Industria
#> 28                                         Explotación Estadística del Directorio Central de Empresas
#> 29             Encuesta sobre Comercio Internacional de Servicios y otras Operaciones Internacionales
#> 30                       Índices de Precios de Exportación y de Importación de Productos Industriales
#> 31                                                                          Distribución de Apellidos
#> 32                                              Contabilidad Nacional Trimestral de España. Base 2008
#> 33                                                            Proyecciones de Población a Corto Plazo
#> 34                                                            Proyecciones de Población a Largo Plazo
#> 35                                                       Índice de Precios de Apartamentos Turísticos
#> 36                                                                       Índice de Precios de Camping
#> 37                                                 Índice de Precios de Alojamientos de Turismo Rural
#> 38                                                 Estadística de Nulidades, Separaciones y Divorcios
#> 39                                                                         Estadística de Migraciones
#> 40                                                                                Cifras de Población
#> 41                                                             Encuesta Industrial Anual de Productos
#> 42                 Índices de Precios de Materiales y Energía e Índices Nacionales de la Mano de Obra
#> 43                                                         Encuesta Cuatrienal de Estructura Salarial
#> 44                                                              Estadística de Sociedades Mercantiles
#> 45                                              Estadística Estructural de Empresas: Sector Servicios
#> 46                                                                        Índice de Precios Hoteleros
#> 47                                                                      Índice de Precios del Trabajo
#> 48                                                                    Encuesta Anual de Coste Laboral
#> 49                                                              Encuesta Anual de Estructura Salarial
#> 50                                                              Encuesta de Condiciones de Vida (ECV)
#> 51                                                               Indicadores de Confianza Empresarial
#> 52                                                           Estadística sobre Transporte Ferroviario
#> 53                                                    Indicadores de Rentabilidad del Sector Hotelero
#> 54                                                                               Tablas de Mortalidad
#> 55                                                                      Flujos de la Población Activa
#> 56                                                                     Estadística de Juzgados de Paz
#> 57                                                                 Estadística de Condenados: Adultos
#> 58                                                                 Estadística de Condenados: Menores
#> 59                                           Estadística de Violencia Doméstica y Violencia de Género
#> 60                                                      Estadística de Movilidad Laboral y Geográfica
#> 61                                  Contabilidad Nacional Trimestral de España: Principales Agregados
#> 62                                                                     Encuesta de Ocupación Hotelera
#> 63                                                   Encuesta de Ocupación en Apartamentos Turísticos
#> 64                                                                  Encuesta de Ocupación en Campings
#> 65                                             Encuesta de Ocupación en Alojamientos de Turismo Rural
#> 66                                Cuentas Trimestrales no Financieras de los Sectores Institucionales
#> 67                                                         Contabilidad Nacional de España. Base 2010
#> 68                                                     Encuesta Coyuntural sobre Stocks y Existencias
#> 69                                                                       Índice de ingresos hoteleros
#> 70                                               Estadística Estructural de Empresas: Sector Comercio
#> 71                                                            Índice de Cifra de Negocios Empresarial
#> 72                                                         Estadística sobre Ejecuciones Hipotecarias
#> 73                                                                 Encuesta de Población Activa (EPA)
#> 74                                                              Estadística de Transporte de Viajeros
#> 75                                                        Encuesta Trimestral de Coste Laboral (ETCL)
#> 76                                                                     MNP Estadística de Matrimonios
#> 77                                                                     MNP Estadística de Nacimientos
#> 78                                                                     MNP Estadística de Defunciones
#> 79                                                          Encuesta de Presupuestos Familiares (EPF)
#> 80                                                                 Encuesta de Ocupación en Albergues
#> 81                                                                        Encuesta de Gasto Turístico
#> 82                                                                Movimientos Turísticos en Fronteras
#> 83                                                            Índice de Garantía de la Competitividad
#> 84                                                                        Mujeres y Hombres en España
#> 85                                                                  Encuesta de turismo de residentes
#> 86                                Estadística de Adquisiciones de Nacionalidad Española de Residentes
#> 87                             Contabilidad nacional anual de España: agregados por rama de actividad
#> 88                                                      Atlas de Distribución de Renta de los Hogares
#> 89                                                                                  Movilidad COVID19
#> 90                                              Estimacion del numero de defunciones semanales (EDeS)
#> 91                                                        Encuesta Cuatrienal de Estructural Salarial
#> 92                                                                     Viviendas turísticas en España
#> 93 Distribución del gasto en destino realizado por los visitantes extranjeros en sus visitas a España
#> 94                                                                  Estimación Mensual de Nacimientos
#> 95                                                   Comercio diario al por menor de grandes empresas
#> 96                                                                  Coyuntura demográfica de empresas
#> 97                                                               Ocupación en alojamientos turísticos
#> 98                                                          Índice de Precios de Vivienda en Alquiler
#> 99                                                      Indicador Multidimensional de Calidad de Vida
#>      Codigo
#> 1        EI
#> 2      ICLA
#> 3      ETDP
#> 4        UA
#> 5       EPC
#> 6       IPS
#> 7       IPV
#> 8      TNOM
#> 9      IPCA
#> 10 CNTR2000
#> 11    EPOBA
#> 12     DPOP
#> 13      ECM
#> 14  EEE:IND
#> 15      IPC
#> 16      IPI
#> 17     IPRI
#> 18      CCM
#> 19      IAS
#> 20      ICM
#> 21      IDB
#> 22     DPOH
#> 23     DPOD
#> 24      EOT
#> 25      HPT
#> 26      IEP
#> 27      ICN
#> 28      DIR
#> 29     ICES
#> 30   IPRX-M
#> 31     APEL
#> 32 CNTR2008
#> 33    EPOBC
#> 34    EPOBL
#> 35     IPAP
#> 36     IPAC
#> 37     IPTR
#> 38     ENSD
#> 39       EM
#> 40       CP
#> 41     EIAP
#> 42      IMM
#> 43   EAES:Q
#> 44       SM
#> 45  EEE:SER
#> 46      IPH
#> 47      IPT
#> 48     EACL
#> 49     EAES
#> 50      ECV
#> 51      ICE
#> 52       TF
#> 53     IRSH
#> 54       TM
#> 55     EFPA
#> 56      EJP
#> 57     CONA
#> 58     CONM
#> 59      VGD
#> 60     EMLG
#> 61 CNTR2010
#> 62      EOH
#> 63     EOAP
#> 64     EOAC
#> 65     EOTR
#> 66   CTNFSI
#> 67      CNE
#> 68     ECSE
#> 69      IIH
#> 70  EEE:COM
#> 71     ICNE
#> 72       EH
#> 73      EPA
#> 74       TV
#> 75     ETCL
#> 76     MNPM
#> 77     MNPN
#> 78     MNPD
#> 79      EPF
#> 80      EOA
#> 81       EG
#> 82       FR
#> 83      IGC
#> 84      MYH
#> 85      ETR
#> 86     ANES
#> 87    CNEAG
#> 88     ADRH
#> 89     MMOV
#> 90     EDES
#> 91   EAES:Q
#> 92      VTE
#> 93     GDVE
#> 94      EMN
#> 95    CDMGE
#> 96    CODEM
#> 97      OAT
#> 98     IPVA
#> 99     IMCV
#>                                                                                                                      Url
#> 1                                                                                                                   <NA>
#> 2                                                                                                                   <NA>
#> 3                                                                                                                   <NA>
#> 4                                                                                                                   <NA>
#> 5                                                                                                                   <NA>
#> 6                                                                                                                   <NA>
#> 7                                                                                                                   <NA>
#> 8   https://www.ine.es/dyngs/INEbase/es/operacion.htm?c=Estadistica_C&cid=1254736177009&menu=ultiDatos&idp=1254734710990
#> 9                                                                                                                   <NA>
#> 10                                                                                                                  <NA>
#> 11                                                                                                                  <NA>
#> 12                                                                                                                  <NA>
#> 13                                                                                                                  <NA>
#> 14                                                                                                                  <NA>
#> 15                                                                                                                  <NA>
#> 16                                                                                                                  <NA>
#> 17                                                                                                                  <NA>
#> 18                                                                                                                  <NA>
#> 19                                                                                                                  <NA>
#> 20                                                                                                                  <NA>
#> 21                                                                                                                  <NA>
#> 22                                                                                                                  <NA>
#> 23                                                                                                                  <NA>
#> 24                                                                                                                  <NA>
#> 25                                                                                                                  <NA>
#> 26                                                                                                                  <NA>
#> 27                                                                                                                  <NA>
#> 28                                                                                                                  <NA>
#> 29                                                                                                                  <NA>
#> 30                                                                                                                  <NA>
#> 31  https://www.ine.es/dyngs/INEbase/es/operacion.htm?c=Estadistica_C&cid=1254736177009&menu=ultiDatos&idp=1254734710990
#> 32                                                                                                                  <NA>
#> 33                                                                                                                  <NA>
#> 34                                                                                                                  <NA>
#> 35                                                                                                                  <NA>
#> 36                                                                                                                  <NA>
#> 37                                                                                                                  <NA>
#> 38                                                                                                                  <NA>
#> 39                                                                                                                  <NA>
#> 40                                                                                                                  <NA>
#> 41                                                                                                                  <NA>
#> 42                                                                                                                  <NA>
#> 43                                                                                                                  <NA>
#> 44                                                                                                                  <NA>
#> 45                                                                                                                  <NA>
#> 46                                                                                                                  <NA>
#> 47                                                                                                                  <NA>
#> 48                                                                                                                  <NA>
#> 49                                                                                                                  <NA>
#> 50                                                                                                                  <NA>
#> 51                                                                                                                  <NA>
#> 52                                                                                                                  <NA>
#> 53                                                                                                                  <NA>
#> 54                                                                                                                  <NA>
#> 55                                                                                                                  <NA>
#> 56                                                                                                                  <NA>
#> 57                                                                                                                  <NA>
#> 58                                                                                                                  <NA>
#> 59                                                                                                                  <NA>
#> 60                                                                                                                  <NA>
#> 61                                                                                                                  <NA>
#> 62                                                                                                                  <NA>
#> 63                                                                                                                  <NA>
#> 64                                                                                                                  <NA>
#> 65                                                                                                                  <NA>
#> 66                                                                                                                  <NA>
#> 67                                                                                                                  <NA>
#> 68                                                                                                                  <NA>
#> 69                                                                                                                  <NA>
#> 70                                                                                                                  <NA>
#> 71                                                                                                                  <NA>
#> 72                                                                                                                  <NA>
#> 73                                                                                                                  <NA>
#> 74                                                                                                                  <NA>
#> 75                                                                                                                  <NA>
#> 76                                                                                                                  <NA>
#> 77                                                                                                                  <NA>
#> 78                                                                                                                  <NA>
#> 79                                                                                                                  <NA>
#> 80                                                                                                                  <NA>
#> 81                                                                                                                  <NA>
#> 82                                                                                                                  <NA>
#> 83 https://www.ine.es/ss/Satellite?c=Page&cid=1259947704944&pagename=ProductosYServicios%2FPYSLayout&L=0&p=1254735893337
#> 84                                                                                                                  <NA>
#> 85                                                                                                                  <NA>
#> 86                                                                                                                  <NA>
#> 87                                                                                                                  <NA>
#> 88                                                          https://www.ine.es/experimental/atlas/experimental_atlas.htm
#> 89                                                         https://www.ine.es/experimental/movilidad/experimental_em.htm
#> 90                                              https://www.ine.es/experimental/defunciones/experimental_defunciones.htm
#> 91                                                                                                                  <NA>
#> 92                                          https://www.ine.es/experimental/viv_turistica/experimental_viv_turistica.htm
#> 93                                                         https://www.ine.es/experimental/gasto_tarjetas/trimestral.htm
#> 94                                              https://www.ine.es/experimental/nacimientos/experimental_nacimientos.htm
#> 95                                                                                https://www.ine.es/experimental/cdmge/
#> 96                                                          https://www.ine.es/experimental/codem/experimental_codem.htm
#> 97                                                                                                                  <NA>
#> 98                                                                                                                  <NA>
#> 99                                          https://www.ine.es/experimental/imcv/experimental_ind_multi_calidad_vida.htm
get_operations(resource = "all")
#>     Id Cod_IOE
#> 1    4   30147
#> 2    6   30211
#> 3    7   30168
#> 4   10   30256
#> 5   13   30219
#> 6   14   30182
#> 7   15   30457
#> 8   16        
#> 9   18   30180
#> 10  20   30013
#> 11  21   30259
#> 12  22   30245
#> 13  23   30417
#> 14  24   30048
#> 15  25   30138
#> 16  26   30050
#> 17  27   30051
#> 18  30        
#> 19  31   30183
#> 20  32   30103
#> 21  33   30264
#> 22  35        
#> 23  36        
#> 24  39   30162
#> 25  40   30149
#> 26  41   30053
#> 27  42   30052
#> 28  43   30203
#> 29  44   30198
#> 30  48   30071
#> 31  49        
#> 32  51   30019
#> 33  52   30269
#> 34  54   30270
#> 35  61   30222
#> 36  62   30223
#> 37  63   30229
#> 38  66   30463
#> 39  71   30277
#> 40  72   30321
#> 41  93   30049
#> 42 104   30062
#> 43 121   30133
#> 44 125   30151
#> 45 130   30177
#> 46 132   30179
#> 47 137   30185
#> 48 139   30188
#> 49 140   30189
#> 50 155   30453
#> 51 163   30199
#> 52 171   30221
#> 53 180   30230
#> 54 197   30271
#> 55 205   30310
#> 56 212   30465
#> 57 213   30466
#> 58 214   30467
#> 59 215   30468
#> 60 234   30209
#> 61 237   30024
#> 62 238   30235
#> 63 239   30236
#> 64 240   30237
#> 65 241   30238
#> 66 246   30026
#> 67 247   30023
#> 68 249   30063
#> 69 254   30220
#> 70 256   30232
#> 71 258   30083
#> 72 259   30153
#> 73 293   30308
#> 74 297   30163
#> 75 303   30187
#> 76 305   30302
#> 77 307   30304
#> 78 309   30306
#> 79 314   30458
#> 80 328   30239
#> 81 329   16029
#> 82 330   16028
#> 83 331        
#> 84 333   30456
#> 85 334   16023
#> 86 336   30279
#> 87 345   30030
#> 88 353        
#> 89 402        
#> 90 406   30324
#> 91 407        
#> 92 410        
#> 93 412        
#> 94 424   30323
#> 95 425        
#> 96 426        
#> 97 429        
#> 98 432        
#> 99 433        
#>                                                                                                Nombre
#> 1                                                        Estadística de Efectos de Comercio Impagados
#> 2                                                                  Índice de Coste Laboral Armonizado
#> 3                                              Estadística de Transmisión de Derechos de la Propiedad
#> 4                                                                                 Indicadores Urbanos
#> 5                                                             Estadística del Procedimiento Concursal
#> 6                                                             Índices de Precios del Sector Servicios
#> 7                                                              Índice de Precios de la Vivienda (IPV)
#> 8                                                                             Distribución de Nombres
#> 9                                                      Índice de Precios de Consumo Armonizado (IPCA)
#> 10                                              Contabilidad Nacional Trimestral de España. Base 2000
#> 11                                                        Estimaciones de la Población Actual (ePOBa)
#> 12           Cifras Oficiales de Población de los Municipios Españoles: Revisión del Padrón Municipal
#> 13                                                Estadística de Defunciones según la Causa de Muerte
#> 14                                             Estadística Estructural de Empresas: Sector Industrial
#> 15                                                                 Índice de Precios de Consumo (IPC)
#> 16                                                                   Índices de Producción Industrial
#> 17                                                                    Índices de Precios Industriales
#> 18                                          Encuesta de coyuntura de comercio al por menor. Base 1994
#> 19                                                      Indicadores de Actividad del Sector Servicios
#> 20                                                                   Índices de Comercio al por Menor
#> 21                                                                   Indicadores Demográficos Básicos
#> 22    Poblaciones de hecho desde 1900 hasta 1991. Cifras oficiales sacadas de los Censos respectivos.
#> 23                 Poblaciones de derecho desde 1986 hasta 1995. Cifras oficiales sacadas del Padrón.
#> 24                                                   Encuesta de Ocupación en Alojamientos Turísticos
#> 25                                                                           Estadística de Hipotecas
#> 26                                                                     Índices de Entradas de Pedidos
#> 27                                                      Índices de Cifras de Negocios en la Industria
#> 28                                         Explotación Estadística del Directorio Central de Empresas
#> 29             Encuesta sobre Comercio Internacional de Servicios y otras Operaciones Internacionales
#> 30                       Índices de Precios de Exportación y de Importación de Productos Industriales
#> 31                                                                          Distribución de Apellidos
#> 32                                              Contabilidad Nacional Trimestral de España. Base 2008
#> 33                                                            Proyecciones de Población a Corto Plazo
#> 34                                                            Proyecciones de Población a Largo Plazo
#> 35                                                       Índice de Precios de Apartamentos Turísticos
#> 36                                                                       Índice de Precios de Camping
#> 37                                                 Índice de Precios de Alojamientos de Turismo Rural
#> 38                                                 Estadística de Nulidades, Separaciones y Divorcios
#> 39                                                                         Estadística de Migraciones
#> 40                                                                                Cifras de Población
#> 41                                                             Encuesta Industrial Anual de Productos
#> 42                 Índices de Precios de Materiales y Energía e Índices Nacionales de la Mano de Obra
#> 43                                                         Encuesta Cuatrienal de Estructura Salarial
#> 44                                                              Estadística de Sociedades Mercantiles
#> 45                                              Estadística Estructural de Empresas: Sector Servicios
#> 46                                                                        Índice de Precios Hoteleros
#> 47                                                                      Índice de Precios del Trabajo
#> 48                                                                    Encuesta Anual de Coste Laboral
#> 49                                                              Encuesta Anual de Estructura Salarial
#> 50                                                              Encuesta de Condiciones de Vida (ECV)
#> 51                                                               Indicadores de Confianza Empresarial
#> 52                                                           Estadística sobre Transporte Ferroviario
#> 53                                                    Indicadores de Rentabilidad del Sector Hotelero
#> 54                                                                               Tablas de Mortalidad
#> 55                                                                      Flujos de la Población Activa
#> 56                                                                     Estadística de Juzgados de Paz
#> 57                                                                 Estadística de Condenados: Adultos
#> 58                                                                 Estadística de Condenados: Menores
#> 59                                           Estadística de Violencia Doméstica y Violencia de Género
#> 60                                                      Estadística de Movilidad Laboral y Geográfica
#> 61                                  Contabilidad Nacional Trimestral de España: Principales Agregados
#> 62                                                                     Encuesta de Ocupación Hotelera
#> 63                                                   Encuesta de Ocupación en Apartamentos Turísticos
#> 64                                                                  Encuesta de Ocupación en Campings
#> 65                                             Encuesta de Ocupación en Alojamientos de Turismo Rural
#> 66                                Cuentas Trimestrales no Financieras de los Sectores Institucionales
#> 67                                                         Contabilidad Nacional de España. Base 2010
#> 68                                                     Encuesta Coyuntural sobre Stocks y Existencias
#> 69                                                                       Índice de ingresos hoteleros
#> 70                                               Estadística Estructural de Empresas: Sector Comercio
#> 71                                                            Índice de Cifra de Negocios Empresarial
#> 72                                                         Estadística sobre Ejecuciones Hipotecarias
#> 73                                                                 Encuesta de Población Activa (EPA)
#> 74                                                              Estadística de Transporte de Viajeros
#> 75                                                        Encuesta Trimestral de Coste Laboral (ETCL)
#> 76                                                                     MNP Estadística de Matrimonios
#> 77                                                                     MNP Estadística de Nacimientos
#> 78                                                                     MNP Estadística de Defunciones
#> 79                                                          Encuesta de Presupuestos Familiares (EPF)
#> 80                                                                 Encuesta de Ocupación en Albergues
#> 81                                                                        Encuesta de Gasto Turístico
#> 82                                                                Movimientos Turísticos en Fronteras
#> 83                                                            Índice de Garantía de la Competitividad
#> 84                                                                        Mujeres y Hombres en España
#> 85                                                                  Encuesta de turismo de residentes
#> 86                                Estadística de Adquisiciones de Nacionalidad Española de Residentes
#> 87                             Contabilidad nacional anual de España: agregados por rama de actividad
#> 88                                                      Atlas de Distribución de Renta de los Hogares
#> 89                                                                                  Movilidad COVID19
#> 90                                              Estimacion del numero de defunciones semanales (EDeS)
#> 91                                                        Encuesta Cuatrienal de Estructural Salarial
#> 92                                                                     Viviendas turísticas en España
#> 93 Distribución del gasto en destino realizado por los visitantes extranjeros en sus visitas a España
#> 94                                                                  Estimación Mensual de Nacimientos
#> 95                                                   Comercio diario al por menor de grandes empresas
#> 96                                                                  Coyuntura demográfica de empresas
#> 97                                                               Ocupación en alojamientos turísticos
#> 98                                                          Índice de Precios de Vivienda en Alquiler
#> 99                                                      Indicador Multidimensional de Calidad de Vida
#>      Codigo
#> 1        EI
#> 2      ICLA
#> 3      ETDP
#> 4        UA
#> 5       EPC
#> 6       IPS
#> 7       IPV
#> 8      TNOM
#> 9      IPCA
#> 10 CNTR2000
#> 11    EPOBA
#> 12     DPOP
#> 13      ECM
#> 14  EEE:IND
#> 15      IPC
#> 16      IPI
#> 17     IPRI
#> 18      CCM
#> 19      IAS
#> 20      ICM
#> 21      IDB
#> 22     DPOH
#> 23     DPOD
#> 24      EOT
#> 25      HPT
#> 26      IEP
#> 27      ICN
#> 28      DIR
#> 29     ICES
#> 30   IPRX-M
#> 31     APEL
#> 32 CNTR2008
#> 33    EPOBC
#> 34    EPOBL
#> 35     IPAP
#> 36     IPAC
#> 37     IPTR
#> 38     ENSD
#> 39       EM
#> 40       CP
#> 41     EIAP
#> 42      IMM
#> 43   EAES:Q
#> 44       SM
#> 45  EEE:SER
#> 46      IPH
#> 47      IPT
#> 48     EACL
#> 49     EAES
#> 50      ECV
#> 51      ICE
#> 52       TF
#> 53     IRSH
#> 54       TM
#> 55     EFPA
#> 56      EJP
#> 57     CONA
#> 58     CONM
#> 59      VGD
#> 60     EMLG
#> 61 CNTR2010
#> 62      EOH
#> 63     EOAP
#> 64     EOAC
#> 65     EOTR
#> 66   CTNFSI
#> 67      CNE
#> 68     ECSE
#> 69      IIH
#> 70  EEE:COM
#> 71     ICNE
#> 72       EH
#> 73      EPA
#> 74       TV
#> 75     ETCL
#> 76     MNPM
#> 77     MNPN
#> 78     MNPD
#> 79      EPF
#> 80      EOA
#> 81       EG
#> 82       FR
#> 83      IGC
#> 84      MYH
#> 85      ETR
#> 86     ANES
#> 87    CNEAG
#> 88     ADRH
#> 89     MMOV
#> 90     EDES
#> 91   EAES:Q
#> 92      VTE
#> 93     GDVE
#> 94      EMN
#> 95    CDMGE
#> 96    CODEM
#> 97      OAT
#> 98     IPVA
#> 99     IMCV
#>                                                                                                                      Url
#> 1                                                                                                                   <NA>
#> 2                                                                                                                   <NA>
#> 3                                                                                                                   <NA>
#> 4                                                                                                                   <NA>
#> 5                                                                                                                   <NA>
#> 6                                                                                                                   <NA>
#> 7                                                                                                                   <NA>
#> 8   https://www.ine.es/dyngs/INEbase/es/operacion.htm?c=Estadistica_C&cid=1254736177009&menu=ultiDatos&idp=1254734710990
#> 9                                                                                                                   <NA>
#> 10                                                                                                                  <NA>
#> 11                                                                                                                  <NA>
#> 12                                                                                                                  <NA>
#> 13                                                                                                                  <NA>
#> 14                                                                                                                  <NA>
#> 15                                                                                                                  <NA>
#> 16                                                                                                                  <NA>
#> 17                                                                                                                  <NA>
#> 18                                                                                                                  <NA>
#> 19                                                                                                                  <NA>
#> 20                                                                                                                  <NA>
#> 21                                                                                                                  <NA>
#> 22                                                                                                                  <NA>
#> 23                                                                                                                  <NA>
#> 24                                                                                                                  <NA>
#> 25                                                                                                                  <NA>
#> 26                                                                                                                  <NA>
#> 27                                                                                                                  <NA>
#> 28                                                                                                                  <NA>
#> 29                                                                                                                  <NA>
#> 30                                                                                                                  <NA>
#> 31  https://www.ine.es/dyngs/INEbase/es/operacion.htm?c=Estadistica_C&cid=1254736177009&menu=ultiDatos&idp=1254734710990
#> 32                                                                                                                  <NA>
#> 33                                                                                                                  <NA>
#> 34                                                                                                                  <NA>
#> 35                                                                                                                  <NA>
#> 36                                                                                                                  <NA>
#> 37                                                                                                                  <NA>
#> 38                                                                                                                  <NA>
#> 39                                                                                                                  <NA>
#> 40                                                                                                                  <NA>
#> 41                                                                                                                  <NA>
#> 42                                                                                                                  <NA>
#> 43                                                                                                                  <NA>
#> 44                                                                                                                  <NA>
#> 45                                                                                                                  <NA>
#> 46                                                                                                                  <NA>
#> 47                                                                                                                  <NA>
#> 48                                                                                                                  <NA>
#> 49                                                                                                                  <NA>
#> 50                                                                                                                  <NA>
#> 51                                                                                                                  <NA>
#> 52                                                                                                                  <NA>
#> 53                                                                                                                  <NA>
#> 54                                                                                                                  <NA>
#> 55                                                                                                                  <NA>
#> 56                                                                                                                  <NA>
#> 57                                                                                                                  <NA>
#> 58                                                                                                                  <NA>
#> 59                                                                                                                  <NA>
#> 60                                                                                                                  <NA>
#> 61                                                                                                                  <NA>
#> 62                                                                                                                  <NA>
#> 63                                                                                                                  <NA>
#> 64                                                                                                                  <NA>
#> 65                                                                                                                  <NA>
#> 66                                                                                                                  <NA>
#> 67                                                                                                                  <NA>
#> 68                                                                                                                  <NA>
#> 69                                                                                                                  <NA>
#> 70                                                                                                                  <NA>
#> 71                                                                                                                  <NA>
#> 72                                                                                                                  <NA>
#> 73                                                                                                                  <NA>
#> 74                                                                                                                  <NA>
#> 75                                                                                                                  <NA>
#> 76                                                                                                                  <NA>
#> 77                                                                                                                  <NA>
#> 78                                                                                                                  <NA>
#> 79                                                                                                                  <NA>
#> 80                                                                                                                  <NA>
#> 81                                                                                                                  <NA>
#> 82                                                                                                                  <NA>
#> 83 https://www.ine.es/ss/Satellite?c=Page&cid=1259947704944&pagename=ProductosYServicios%2FPYSLayout&L=0&p=1254735893337
#> 84                                                                                                                  <NA>
#> 85                                                                                                                  <NA>
#> 86                                                                                                                  <NA>
#> 87                                                                                                                  <NA>
#> 88                                                          https://www.ine.es/experimental/atlas/experimental_atlas.htm
#> 89                                                         https://www.ine.es/experimental/movilidad/experimental_em.htm
#> 90                                              https://www.ine.es/experimental/defunciones/experimental_defunciones.htm
#> 91                                                                                                                  <NA>
#> 92                                          https://www.ine.es/experimental/viv_turistica/experimental_viv_turistica.htm
#> 93                                                         https://www.ine.es/experimental/gasto_tarjetas/trimestral.htm
#> 94                                              https://www.ine.es/experimental/nacimientos/experimental_nacimientos.htm
#> 95                                                                                https://www.ine.es/experimental/cdmge/
#> 96                                                          https://www.ine.es/experimental/codem/experimental_codem.htm
#> 97                                                                                                                  <NA>
#> 98                                                                                                                  <NA>
#> 99                                          https://www.ine.es/experimental/imcv/experimental_ind_multi_calidad_vida.htm
get_operations(resource = "all", help = TRUE)
#> Available params for resource = "all" are: 
#> - lang
#> Example (basic): get_operations()
#> Example (extended): get_operations(resource = "all", lang = "ES")
if (FALSE) {
get_operations("IPC", resource = "metadata")
get_operations("IPC", resource = "variables_values")
get_operations(resource = "by_granularity", geographical_granularity = "PROV")
}

Site built with pkgdown 2.0.4.

Template rostemplate by dieghernan, based on Bootstrapious.