Get values

This function returns values from a variable or variable operation

Usage

get_values(
  code = NULL,
  resource = "all",
  operation = NULL,
  help = FALSE,
  det = 0,
  ioe = FALSE,
  lang = "ES"
)

Arguments

code

(int) variable indentificator

resource

(string) resource to access, by default resource = "all" to get serie metadata. Possible values are all or variable_operation

operation

(string) operation indentificator

help

(boolean) type any value for resource param and type help = TRUE to see params available for this resource.

det

(int) det = 1 to see the detail of the variable to which it belongs, det = 0 by default

ioe

(boolean) TRUE if code is in format IO30138, and FALSE by default

lang

(string) language used to obtain information

Examples

get_values(115)
#>        Id Fk_Variable                       Nombre Codigo
#> 1       2         115                  Araba/Álava     01
#> 2       3         115                     Albacete     02
#> 3       4         115             Alicante/Alacant     03
#> 4       5         115                      Almería     04
#> 5       6         115                        Ávila     05
#> 6       7         115                      Badajoz     06
#> 7       8         115               Balears, Illes     07
#> 8       9         115                    Barcelona     08
#> 9      10         115                       Burgos     09
#> 10     11         115                      Cáceres     10
#> 11     12         115                        Cádiz     11
#> 12     13         115           Castellón/Castelló     12
#> 13     14         115                  Ciudad Real     13
#> 14     15         115                      Córdoba     14
#> 15     16         115                    Coruña, A     15
#> 16     17         115                       Cuenca     16
#> 17     18         115                       Girona     17
#> 18     19         115                      Granada     18
#> 19     20         115                  Guadalajara     19
#> 20     21         115                     Gipuzkoa     20
#> 21     22         115                       Huelva     21
#> 22     23         115                       Huesca     22
#> 23     24         115                         Jaén     23
#> 24     25         115                         León     24
#> 25     26         115                       Lleida     25
#> 26     27         115                    Rioja, La     26
#> 27     28         115                         Lugo     27
#> 28     29         115                       Madrid     28
#> 29     30         115                       Málaga     29
#> 30     31         115                       Murcia     30
#> 31     32         115                      Navarra     31
#> 32     33         115                     Asturias     33
#> 33     34         115                     Palencia     34
#> 34     35         115                  Palmas, Las     35
#> 35     36         115                   Pontevedra     36
#> 36     37         115                    Salamanca     37
#> 37     38         115       Santa Cruz de Tenerife     38
#> 38     39         115                    Cantabria     39
#> 39     40         115                      Segovia     40
#> 40     41         115                      Sevilla     41
#> 41     42         115                        Soria     42
#> 42     43         115                    Tarragona     43
#> 43     44         115                       Teruel     44
#> 44     45         115                       Toledo     45
#> 45     46         115            Valencia/València     46
#> 46     47         115                   Valladolid     47
#> 47     48         115                      Bizkaia     48
#> 48     49         115                       Zamora     49
#> 49     50         115                     Zaragoza     50
#> 50     51         115                        Ceuta     51
#> 51     52         115                      Melilla     52
#> 52     53         115                      Ourense     32
#> 53     68         115                   Extranjero       
#> 54  15299         115     Total Ciudades con Metro       
#> 55  16473         115               Total Nacional     00
#> 56 274508         115 Total capitales de provincia  00000
#> 57 285494         115  Otras Comunidades Autónomas       
#> 58 291907         115         Peninsula y Baleares       
get_values(resource = "all", help = TRUE)
#> Available params for resource = "all" are: 
#> - code (variable id)
#> - det
#> - lang
#> Example (basic): get_values(115)
#> Example (extended): get_values(115, resource = "all", det = 0, lang = "ES")
get_values(762, operation = "IPC", resource = "variable_operation")
#>        Id Fk_Variable
#> 1  304092         762
#> 2  304093         762
#> 3  304094         762
#> 4  304095         762
#> 5  304096         762
#> 6  304097         762
#> 7  304098         762
#> 8  304099         762
#> 9  304100         762
#> 10 304101         762
#> 11 304102         762
#> 12 304103         762
#> 13 304104         762
#>                                                                                Nombre
#> 1                                                                      Índice general
#> 2                                                  Alimentos y bebidas no alcohólicas
#> 3                                                        Bebidas alcohólicas y tabaco
#> 4                                                                   Vestido y calzado
#> 5                              Vivienda, agua, electricidad, gas y otros combustibles
#> 6  Muebles, artículos del hogar y artículos para el mantenimiento corriente del hogar
#> 7                                                                             Sanidad
#> 8                                                                          Transporte
#> 9                                                                      Comunicaciones
#> 10                                                                     Ocio y cultura
#> 11                                                                          Enseñanza
#> 12                                                             Restaurantes y hoteles
#> 13                                                          Otros bienes y servicios 
#>    Codigo
#> 1      00
#> 2      01
#> 3      02
#> 4      03
#> 5      04
#> 6      05
#> 7      06
#> 8      07
#> 9      08
#> 10     09
#> 11     10
#> 12     11
#> 13     12

Site built with pkgdown 2.0.4.

Template rostemplate by dieghernan, based on Bootstrapious.